Move all sources into /src
