+++ src/__specs__/catchErrors.spec.js
