+++ src/__specs__/log.spec.js
