+++ src/electron-log.d.ts
