+++ src/transports/file/__specs__/findLogPath.spec.js
