+++ src/transports/file/__specs__/getAppName.spec.js
