+++ src/transports/file/__specs__/index.spec.js
