+++ src/transports/rendererConsole.js
