+++ test-projects/webpack/src/main.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var electron = require('electron');
-var log      = require('electron-log');
+var log      = require('src/electron-log');
 var path     = require('path');
 
 var app = electron.app;
